Add heavy array forecast project
