+++ examples/PerfIssueLib/HeavyArrayForecast/Readme.md
@@ -1,0 +1,5 @@
+# Heavy Weather Forcast
+
+Generate weather forecast for multiple times, unnecessarily realize the list and return the result once.
+
+Use this for a performance issue demo.
