+++ examples/QuickStart3_1/global.json
@@ -1,5 +1,6 @@
 {
   "sdk": {
-    "version": "3.1.407"
+    "version": "3.1.407",
+    "rollForward": "latestMajor"
   }
 }
