+++ examples/QuickStart3_1/test.http
@@ -1,0 +1,1 @@
+GET http://localhost:5000/weatherforecast/
